refactor: separate spi driver from mmc
